backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pymongo==4.8.0
firebase-admin==6.5.0
pydantic-settings==2.5.2
python-dotenv==1.0.1
email-validator==2.2.0
